Move all code to FlyAnimation
